+++ tests/server-function/cypress/e2e/server-function.cy.ts
@@ -1,0 +1,6 @@
+describe("server-function", () => {
+  it("should isServer false on the client and true in the server function", () => {
+    cy.visit("/");
+    cy.get("#server-fn-test").contains(`{"client":false,"serverFn":true}`);
+  })
+});